requirements-freeze.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
packaging==23.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.2
PySide6-Essentials==6.4.3
pyxdg==0.28
qt-material==2.14
regex==2023.3.23
requests==2.28.2
shiboken6==6.4.3
show-in-file-manager==1.1.4
tiktoken==0.3.3
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
